refactor(dashboard): move the embed hosts warning to the dashboard home
